requirements.txt
attrs==19.3.0
bcrypt==3.1.7
bit-array==0.1.0
bitarray==1.2.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cryptography==2.8
Cython==0.29.16
findspark==1.3.0
Fiona==1.8.13
idna==2.9
impyla==0.15.0
munch==2.5.0
numpy==1.18.1
pandas==0.25.3
paramiko==2.7.1
ply==3.11
psycopg2-binary==2.8.4
py4j==0.10.7
pycparser==2.20
PyNaCl==1.3.0
pyodbc==4.0.27
pyproj==2.4.2.post1
pysftp==0.2.9
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
sasl==0.2.1
Shapely==1.7.0
six==1.12.0
snakebite-py3==3.0.5
SQLAlchemy==1.3.10
thrift==0.11.0
thrift-sasl==0.2.1
thriftpy==0.3.9
thriftpy2==0.4.0
tqdm==4.45.0
urllib3==1.25.8
WebHDFS==0.2.0
